+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Unreleased
 - Run migrations during install to avoid missing tables after database removal
 - Add --satellite flag to install script
 - Add --control flag to install script to create Control site
+- Add --terminal flag to install script
 - Add msg app for LCD/Windows notifications
 - Remove environment sigils integration
 - Show full date and timezone in admin clock tooltip